+++ RELEASE-NOTES.md
@@ -2,6 +2,8 @@
 3.6
 ---
 
+* support CupsGetDevices operation
+* Update `printer-is-accepting-jobs` attribute after `CupsAcceptJobs` or `CupsRejectJobs` operation
 * Upgraded to Gradle 9.3.0
 * Upgraded to Kotlin Gradle Plugin 2.2.21
 * Upgraded to Kotlin 1.9